perofrmance improvements, new data and upload of data extractor
I've updated the map to 4/26. The `prune` function is much faster thanks to the removal of some unneccesary code. I've also added to the repository the Python script for extracting JSON data from the Excel spreadsheet, in case anyone finds a better way to do it.

Optimizations, bug fixes, and feature requests are greatly appreciated. That's it for today, and I'll see you in the next one.